+++ antlr/test.sh
@@ -5,5 +5,12 @@
 # ensure to use setup.bash first
 # call like this: . test.sh
 
-grun Grammar1 edge -gui -tokens -trace test_input.txt
-# grun Grammar1 node -tokens -trace test_input.txt
+# grun StateSmithLabelGrammar edge -gui -tokens -trace ss_label_test_input.txt
+grun StateSmithLabelGrammar edge -tokens -trace ss_label_test_input.txt
+# grun StateSmithLabelGrammar node -gui -tokens -trace ss_label_test_input.txt
+# grun StateSmithLabelGrammar node -tokens -trace ss_label_test_input.txt > /dev/null
+
+
+# grun PlantUML diagram -gui -tokens -trace puml_test_input_1.puml
+# if you are looking for an error, this will show only stderr
+# grun PlantUML diagram -tokens -trace puml_test_input_1.puml > /dev/null